Rename api_server to http_api
